refactor(kotlin-dsl): 코틀린 DSL 리뷰어님(한솔님) 피드백 반영
- Test에 모든 객체들을 두었던 것을 다른 파일들로 분리함
- @DslMaker 를 통해 외부 리시버를 사용하지 못하게 막기
- @DisplayName 을 통해 테스트에 대한 정보 명시
