+++ src/main/kotlin/study/Dsl.kt
@@ -1,0 +1,8 @@
+package study
+
+@DslMarker
+annotation class DslScope
+
+fun introduce(block: PersonBuilder.() -> Unit): Person {
+    return PersonBuilder().apply(block).build()
+}
