+++ src/main/kotlin/study/Language.kt
@@ -1,0 +1,14 @@
+package study
+
+data class Language(val name: String, val level: Int)
+
+@DslScope
+class LanguagesBuilder {
+    private val languages = mutableListOf<Language>()
+
+    infix fun String.level(level: Int) {
+        languages.add(Language(this, level))
+    }
+
+    fun build(): List<Language> = languages
+}
